Everything works, set default server as app engine
