--- chrome-plugin/src/options/options.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head><title>Legal Leaf Options</title></head>
-<body>
-
-<label>
-  Developer Mode
-  <input type="checkbox" id="developer_mode">
-</label>
-<label>
-    Protocol
-    <input type="text" id="protocol">
-</label>
-<label>
-    IP Address
-    <input type="text" id="server_ip">
-</label>
-<label>
-    Port Number
-    <input type="text" id="port">
-</label>
-<label>
-    Port Number
-    <input type="text" id="path">
-</label>
-
-<div id="status"></div>
-<button id="save">Save</button>
-
-<script src="options.js"></script>
-</body>
-</html>
