+++ chrome_plugin/_locales/.DS_Store
