+++ chrome_plugin/_locales/en/messages.json
