+++ chrome_plugin/icons/.DS_Store
