+++ chrome_plugin/manifest.json
