+++ chrome_plugin/src/bg/.DS_Store
