+++ chrome_plugin/src/browser_action/browser_action.js
