+++ chrome_plugin/src/options/options.js
@@ -24,7 +24,6 @@ function save_options() {
 // Restores select box and checkbox state using the preferences
 // stored in chrome.storage.
 function restore_options() {
-    // Use default value color = 'red' and likesColor = true.
     chrome.storage.sync.get({
         developerMode: false,
         protocol: "https://",
